fix(board): land rule E's fallout — FSM pin, moved patch target, flat-core home (#4152)
Three gates the rule-E commit left red, all consequences of that diff rather
than new behaviour:

- `test_ticket_structure`'s FSM graph pin did not carry `reopen`'s new
  `delivered` source. Pinned, with the edge named beside #3327's.
- `test_ticket_completion` patched `backends.loader.get_code_host_for_url`,
  but the scanner now reaches the resolver through `backends.issue_reads`,
  which binds it at import — so the patch silently no-opped. Retargeted at the
  module that binds it.
- `issue_reopen.py` sat as a new flat `core/` leaf and tripped the
  no-flat-core-regrowth ratchet. It is a forge-payload classifier, not a core
  concern: its sibling in kind (`pr_open_state_from_payload`) already lives
  under `backends/`, and its only consumer is `backends/issue_reads.py`. Moved
  to `backends/issue_reopen.py` (tests mirrored) rather than raising the
  ceiling — `tach check` clean.

Open questions & assumptions

- assumed: the ratchet wanted a move, not a raise. Its own message offers both;
  the leaf has no core consumer and an exact backends precedent, so raising the
  ceiling would have banked a leaf in the wrong package.
- assumed: `test_ticket_completion`'s six failures were the stale patch target
  and nothing else. Retargeting alone turns all six green.
- open: 13 further failures in the affected lane reproduce identically on
  `origin/main` in the main clone — live `gh` credentials, a live control DB,
  and docker in the test environment. Environment-dependent, not this branch.
