+++ tests/teatree_loop/test_ticket_completion.py
@@ -103,9 +103,10 @@ class TicketCompletionScannerTests(TestCase):
 
     def _patch_host(self, host: _Host):
         # The scanner resolves the host through the shared ``issue_is_done`` seam
-        # in ``teatree.backends.loader``, so patch the host resolver there.
+        # in ``teatree.backends.issue_reads``, which binds the resolver at import,
+        # so patching it on ``backends.loader`` would silently no-op.
         return patch(
-            "teatree.backends.loader.get_code_host_for_url",
+            "teatree.backends.issue_reads.get_code_host_for_url",
             return_value=host,
         )
